IncognitoGPT/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openai==1.14.3
sqlite-utils==3.36
cryptography==42.0.5
python-dotenv==1.0.1
